Proof of concept: a working prototype with no templates
This commit adds a prototype blog that loads markdown files from the /posts
folder, lists them on the homepage in order of date modified (most recent
first), and parses markdown files into HTML when viewing a blog post.

Needs some serious refactoring to make it tidy, and the addition of mustache
templates to make it look good.