Adds initial Backbone framework.
Adds a router to handle the homepage and blog views, plus anchor click
handling in Backbone. Next up: blog posts as models.

Modernizr is also loaded, though we may well need to move the requirejs
library to the head tag, load modernizr ASAP and then wrap the rest of the
code in a document load event. Depends on IE.